requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.2
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
# Django core
Django==5.0.6

# Django REST Framework
djangorestframework==3.15.1

# CORS (para permitir acceso desde el frontend)
django-cors-headers==4.3.1

# Pillow (si manejas imágenes, casi siempre es buena idea tenerlo)
Pillow==10.3.0

# Si en algún momento usas PostgreSQL (puedes dejarlo comentado si todavía no)
# psycopg2-binary==2.9.9

# (Opcional si usas JWT en el futuro)
# djangorestframework-simplejwt==5.3.1

# Asgi (manejo de ASGI que ya viene indirectamente con Django 5)
asgiref==3.7.2
